+++ src/mathtools.h
@@ -931,7 +931,7 @@ template <class objecttype>
 void random_perm ( std::vector<objecttype> &s)
 {
 	int len=s.size();
-	for ( size_t i=0; i<len-1; i++ ) {
+	for ( int i=0; i<len-1; i++ ) {
 		int j = i+myrand() % ( len-i );
 		std::swap ( s[i], s[j] );
 	}
